+++ src/test-embedding/java/org/truffleruby/test/embedding/JSR223InteropTest.java
@@ -12,7 +12,6 @@ package org.truffleruby.test.embedding;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import java.util.List;
 
@@ -20,18 +19,23 @@ import javax.script.Bindings;
 import javax.script.Compilable;
 import javax.script.CompiledScript;
 import javax.script.Invocable;
+import javax.script.ScriptContext;
 import javax.script.ScriptEngine;
-import javax.script.ScriptEngineManager;
 import javax.script.ScriptException;
 
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 import org.truffleruby.test.embedding.fixtures.FluidForce;
-import org.truffleruby.services.scriptengine.TruffleRubyScriptEngineFactory;
 
 public class JSR223InteropTest {
 
     private ScriptEngine scriptEngine = null;
+
+    @Before
+    public void before() {
+        scriptEngine = new TruffleRubyEngineFactory().getScriptEngine();
+    }
 
     @After
     public void after() {
@@ -51,49 +55,29 @@ public class JSR223InteropTest {
 
     @Test
     public void testVersion() {
-        assertNotNull(new TruffleRubyScriptEngineFactory().getEngineVersion());
+        assertNotNull(new TruffleRubyEngineFactory().getEngineVersion());
     }
 
     @Test
     public void testCreateEngine() throws ScriptException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         assertEquals(42, scriptEngine.eval("6 * 7"));
     }
 
     @Test
-    public void testNoPermissionsByDefault() {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
-        try {
-            scriptEngine.eval("Process.pid");
-            fail("should have thrown");
-        } catch (ScriptException scriptException) {
-            assertEquals("org.graalvm.polyglot.PolyglotException: native access is not allowed",
-                    scriptException.getMessage());
-        }
-    }
-
-    @Test
     public void testAllAccess() throws ScriptException {
-        scriptEngine = new TruffleRubyScriptEngineFactory().getScriptEngine(true);
         assertTrue(scriptEngine.eval("Process.pid") instanceof Integer);
     }
 
     @Test
     public void testParameters() throws ScriptException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
-        final Bindings bindings = scriptEngine.createBindings();
+        final Bindings bindings = scriptEngine.getBindings(ScriptContext.ENGINE_SCOPE);
         bindings.put("a", 14);
         bindings.put("b", 2);
-        assertEquals(16, scriptEngine.eval("a + b", bindings));
+        assertEquals(16, scriptEngine.eval("Truffle::Boot::INTERACTIVE_BINDING.eval('a + b')"));
     }
 
     @Test
     public void testCallingMethods() throws ScriptException, NoSuchMethodException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         assertEquals(
                 0.909,
                 (double) ((Invocable) scriptEngine).invokeMethod(scriptEngine.eval("Math"), "sin", 2),
@@ -102,8 +86,6 @@ public class JSR223InteropTest {
 
     @Test
     public void testCreatingObjects() throws ScriptException, NoSuchMethodException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         final Object time = ((Invocable) scriptEngine).invokeMethod(scriptEngine.eval("Time"), "new", 2021, 3, 18);
         final Object year = ((Invocable) scriptEngine).invokeMethod(time, "year");
         assertEquals(2021, year);
@@ -112,15 +94,11 @@ public class JSR223InteropTest {
     @SuppressWarnings("unchecked")
     @Test
     public void testAccessingArrays() throws ScriptException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         assertEquals(4, ((List<Object>) scriptEngine.eval("[3, 4, 5]")).get(1));
     }
 
     @Test
     public void testAccessingHashes() throws ScriptException, NoSuchMethodException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         assertEquals(
                 4,
                 (int) ((Invocable) scriptEngine).invokeMethod(
@@ -131,8 +109,6 @@ public class JSR223InteropTest {
 
     @Test
     public void testImplementInterface() throws ScriptException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         final FluidForce fluidForce = ((Invocable) scriptEngine)
                 .getInterface(scriptEngine.eval(FluidForce.RUBY_SOURCE), FluidForce.class);
         assertEquals(5587.008375144088, fluidForce.getFluidForce(2.0, 3.0, 6.0), 0.01);
@@ -140,8 +116,6 @@ public class JSR223InteropTest {
 
     @Test
     public void testParseOnceRunMany() throws ScriptException {
-        final ScriptEngineManager m = new ScriptEngineManager();
-        scriptEngine = m.getEngineByName("ruby");
         final CompiledScript compiled = ((Compilable) scriptEngine).compile("14");
         assertEquals(14, compiled.eval());
     }